[refactor] 커뮤니티 상세 모달창 - 작성자 닉네임 가져오기
